+++ lib/flutter_gyw.dart
@@ -1,9 +1,11 @@
+/// This library provides an interface between a Flutter application and
+/// Get Your Way devices like the aRdent smart glasses.
 library flutter_gyw;
 
 export 'src/bt_device.dart' show BTDevice;
 export 'src/bt_manager.dart' show BTManager;
 export 'src/drawings.dart' show Drawing, TextDrawing, WhiteScreen, IconDrawing;
 export 'src/exceptions.dart' show GYWException, GYWStatusException;
-export 'src/fonts.dart' show GYWFonts;
-export 'src/icons.dart' show GYWIcons;
+export 'src/fonts.dart' show GYWFont, GYWFonts;
+export 'src/icons.dart' show GYWIcon, GYWIcons;
 export 'src/screen.dart' show GYWScreenParameters;