+++ lib/src/exceptions.dart
@@ -1,8 +1,14 @@
 /// This type of exception is triggered by this package in case of error
 class GYWException implements Exception {
+  /// String describing the cause of the error
   final String cause;
 
   const GYWException(this.cause);
+
+  @override
+  String toString() {
+    return "$runtimeType: $cause";
+  }
 }
 
 /// This type of exception is triggered when a method is triggered while an