+++ lib/src/fonts.dart
@@ -3,6 +3,9 @@ class GYWFont {
   /// Index value transmitted to the device
   final int index;
 
+  /// Internal name of the font
+  final String name;
+
   /// Width (in pixels) of a character
   final int width;
 
@@ -13,6 +16,7 @@ class GYWFont {
   final bool bold;
 
   const GYWFont(
+    this.name,
     this.index, [
     this.width = 12,
     this.height = 12,
@@ -22,8 +26,10 @@ class GYWFont {
 
 /// Fonts supported by aRdent devices
 class GYWFonts {
-  static const basic = GYWFont(0, 14, 14, false);
-  static const title = GYWFont(1, 18, 18, true);
-  static const bigBasic = GYWFont(2, 30, 60, false);
-  static const bigTitle = GYWFont(3, 50, 90, true);
+  static const basic = GYWFont("basic", 0, 14, 22, false);
+  static const title = GYWFont("title", 1, 20, 31, true);
+  static const bigBasic = GYWFont("big_basic", 2, 29, 47, false);
+  static const bigTitle = GYWFont("big_title", 3, 42, 71, true);
+
+  static const values = [basic, title, bigBasic, bigTitle];
 }